requirements copy.txt
asgiref==3.3.4
autopep8==1.5.6
certifi==2020.12.5
chardet==4.0.0
Django==3.2
django-3-jet==1.0.7
django-input-mask==2.0.2
django-localflavor==3.0.1
django-widget-tweaks==1.4.8
google-api-python-client==1.4.1
gunicorn==20.1.0
httplib2==0.19.1
idna==2.8
oauth2client==4.1.3
Pillow==8.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
Pygments==2.4.2
pyparsing==2.4.7
python-stdnum==1.16
pytz==2021.1
requests==2.25.1
rsa==4.7.2
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.4
whitenoise==5.2.0
django-heroku==0.3.1 
